+++ operator/tests/e2e/client_cluster.go
@@ -10,6 +10,7 @@ import (
 	"io"
 	"os"
 	"slices"
+	"strings"
 	"sync"
 	"time"
 
@@ -26,6 +27,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/client-go/kubernetes"
@@ -499,15 +501,40 @@ func (cc *clientCluster) verifyPodCounts(ctx context.Context) {
 	Expect(len(targets.Items)).To(Equal(int(cc.cluster.GetTargetSize())))
 }
 
-func (cc *clientCluster) hasPendingTargetPods(ctx context.Context) bool {
-	podList, err := cc.k8sClient.ListPods(ctx, cc.cluster, target.BasicLabels(cc.cluster))
+func (cc *clientCluster) hasPendingTargetPod(ctx context.Context, ordinal int) bool {
+	pod, err := cc.k8sClient.GetPod(ctx, types.NamespacedName{
+		Name:      fmt.Sprintf("%s-%d", target.StatefulSetNSName(cc.cluster).Name, ordinal),
+		Namespace: cc.cluster.Namespace,
+	})
+	if err != nil {
+		return false
+	}
+	return pod.Status.Phase == corev1.PodPending
+}
+
+func (cc *clientCluster) makeTargetUnschedulable(ctx context.Context, ordinal int) {
+	pod := &corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      fmt.Sprintf("%s-%d", target.StatefulSetNSName(cc.cluster).Name, ordinal),
+			Namespace: cc.cluster.Namespace,
+		},
+	}
+	_, err := cc.k8sClient.DeleteResIfExistsWithGracePeriod(ctx, pod, 0)
 	Expect(err).To(BeNil())
-	for i := range podList.Items {
-		if podList.Items[i].Status.Phase == corev1.PodPending {
-			return true
+
+	targetSuffix := fmt.Sprintf("target-%d", ordinal)
+	var pvs []*corev1.PersistentVolume
+	for _, pv := range cc.pvs {
+		if pv.Spec.ClaimRef != nil && strings.HasSuffix(pv.Spec.ClaimRef.Name, targetSuffix) {
+			pvs = append(pvs, pv)
 		}
 	}
-	return false
+	Expect(pvs).NotTo(BeEmpty())
+	tutils.DestroyPV(ctx, cc.k8sClient, pvs)
+
+	Eventually(func(ctx context.Context) bool {
+		return cc.hasPendingTargetPod(ctx, ordinal)
+	}, clusterUpdateTimeout, clusterUpdateInterval).WithContext(ctx).Should(BeTrue())
 }
 
 // Print logs from all pods in this cluster